+++ www/lib/externalaccess.lib.php
@@ -481,9 +481,9 @@ function externalAccessObjectAutoLoad($objecttype, &$db)
 		$module = 'expedition_bon';
 	}
 	elseif ($objecttype == 'delivery') {
-		$classpath = 'livraison/class';
-		$myobject = 'livraison';
-		$module = 'livraison_bon';
+		$classpath = 'delivery/class';
+		$myobject = 'delivery';
+		$module = 'delivery_bon';
 	}
 	elseif ($objecttype == 'contract') {
 		$classpath = 'contrat/class';